Generalize the protocol integration test to extract common code
